Refactor ResponseUtils
